cgi: Move handling of user-submitted form data into own function
handle_submit() is now used to process all non-OpenID POST fields before
any other handler gets to them.  It returns an action  string indicating what
actions should be taken but subsequent handlers.

Current valid actions are:
  - authenticate
  - cancel
  - ask_again

ask_again is not yet handled, and more refactoring still needs to be
done to move all of the flow to the new schema.