poit was NOT working with:   * http://px2owffng8.tal.ki/   * http://test-id.org/ (OP AX fetch)
Seems to be due to missing openid links.
Based on http://wiki.guruj.net/Clamshell!Documentation

A HTML file with that content (and the correct URL) works.
As does injecting the title links in this change.
NOTE http://px2owffng8.tal.ki/ does NOT know what the user
name is (shows unknown), however using a html file with the
delegate links shows the OpenID URL.